Junit test for CompanyName, Item, Quote, QuoteList, QuotelyState

+++ src/main/java/seedu/quotely/data/Item.java
@@ -16,7 +16,7 @@ public class Item {
     }
 
     /**
-     * Unused getters and setters for future versions
+     * Unused getters and setters for future use
      */
 
     public void setItemName(String itemName) {
@@ -27,7 +27,7 @@ public class Item {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
